chore: update dependencies and refactor example projects
- Bumped versions of `@gqloom/core`, `@gqloom/valibot`, and `@gqloom/drizzle` to latest in `pnpm-lock.yaml` and example `package.json` files.
- Refactored TypeScript configuration to use `NodeNext` module resolution.
- Updated resolvers in the cattery examples to utilize the new database query methods and removed the deprecated user service.
- Added a new `hello` resolver to demonstrate basic query functionality.
- Enhanced context handling in resolvers to improve user authentication logic.
