+++ examples/cattery-valibot/src/index.ts
@@ -4,7 +4,7 @@ import { ValibotWeaver } from "@gqloom/valibot"
 import { createYoga } from "graphql-yoga"
 import { resolvers } from "./resolvers"
 
-const schema = weave(ValibotWeaver, ...resolvers)
+const schema = weave(asyncContextProvider, ValibotWeaver, ...resolvers)
 
 const yoga = createYoga({ schema })
 createServer(yoga).listen(4000, () => {
@@ -13,6 +13,7 @@ createServer(yoga).listen(4000, () => {
 
 import * as fs from "fs"
 import * as path from "path"
+import { asyncContextProvider } from "@gqloom/core/context"
 import { printSchema } from "graphql"
 if (process.env.NODE_ENV !== "production") {
   fs.writeFileSync(
