+++ examples/cattery-valibot/src/resolvers/hello.ts
@@ -1,0 +1,8 @@
+import { query, resolver } from "@gqloom/core"
+import * as v from "valibot"
+
+export const helloResolver = resolver({
+  hello: query(v.string())
+    .input({ name: v.nullish(v.string(), "World") })
+    .resolve(({ name }) => `Hello ${name}!`),
+})
