--- examples/cattery-valibot/src/services/index.ts
@@ -1,1 +1,0 @@
-export * as userService from "./user"
